refactor code and write more tests
