+++ app/api/v1/models.py
@@ -1,0 +1,5 @@
+questions=[]
+meetups=[]
+users=[]
+
+
